requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.2
openai==1.59.9
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.5
pydantic_core==2.27.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
